// src/api/vcci-hcm/models/organization.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VCCI
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */

export interface Organization {
  name?: string;
  avatar?: string;
  is_premium?: boolean;
  created_at?: string;
  created_by?: string;
  updated_at?: string;
  updated_by?: string;
  ref_code?: string;
  tax_code?: string;
  org_categories?: string[];
  website?: string;
  address?: string;
  description?: string;
  type?: string;
  banner?: string;
  welcome_media?: string;
  certificates?: string[];
  org_email?: string;
  org_number?: string;
  org_representative?: string;
  org_status_id?: string;
  draft_changes?: string;
  is_hidden?: boolean;
  org_link?: string;
  club_discount?: string;
  welcome_content?: string;
  time_club_discount?: string;
  contact_info?: string;
  sort_order?: number;
  id?: string;
  province?: string[];
  english_name?: string;
  abbreviation?: string;
  phone?: string;
  fax?: string;
  type_of_ownership?: string;
  establish_decision_no?: string;
  establish_decision_time?: string;
  establish_decision_by?: string;
  total_member?: number;
  activities_area?: string;
  total_employee?: number;
  office_address?: string;
  bussiness_regis_no?: string;
  bussiness_regis_time?: string;
  bussiness_regis_by?: string;
  bussiness_regis_img?: string;
  org_representative_office?: string;
  registered_capital?: number;
  total_asset?: number;
  previous_revenue?: number;
  membership_fee?: number;
  social_media_info?: string;
  question?: string;
  signed_file?: string;
}