// src/api/vcci-news/models/video.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */

export interface Video {
  id: string;
  name: string;
  url: string;
  created_at: string;
  /** @nullable */
  created_by?: string | null;
  updated_at: string;
  /** @nullable */
  updated_by?: string | null;
}